gayeon_mulitagent/web_release/requirements.txt
# Web Release 전용 requirements

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Streamlit
streamlit>=1.28.0
plotly>=5.17.0

# Existing dependencies
google-genai>=0.2.0
langchain>=0.1.0
langchain-community>=0.1.0
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
